# Make Ref.read Ref.write instructions